feat #2: 25.02.05 회의 내용 반영
