--- dogether/Data/Model/AppleLoginRequest.swift
@@ -1,14 +1,0 @@
-//
-//  AppleLoginRequest.swift
-//  dogether
-//
-//  Created by seungyooooong on 2/4/25.
-//
-
-import Foundation
-
-// TODO: 추후 수정
-struct AppleLoginRequest: Encodable {
-    let name: String? = "김영재"
-    let idToken: String = "idTokenidTokenidToken"
-}
