--- dogether/Data/Model/CertifyTodoRequest.swift
@@ -1,14 +1,0 @@
-//
-//  CertifyTodoRequest.swift
-//  dogether
-//
-//  Created by seungyooooong on 2/4/25.
-//
-
-import Foundation
-
-// TODO: 추후 수정
-struct CertifyTodoRequest: Encodable {
-    let content: String = "이 노력, 땀 그 모든것이 내 노력의 증거입니다. 양심 있으면 인정 누르시죠."
-    let mediaUrls: [String] = ["https://dogether-bucket-dev.s3.ap-northeast-2.amazonaws.com/daily-todo-proof-media/mock/e1.png", "https://dogether-bucket-dev.s3.ap-northeast-2.amazonaws.com/daily-todo-proof-media/mock/e2.png"]
-}
