+++ dogether/Data/Network/Request/CertifyTodoRequest.swift
@@ -1,0 +1,13 @@
+//
+//  CertifyTodoRequest.swift
+//  dogether
+//
+//  Created by seungyooooong on 2/4/25.
+//
+
+import Foundation
+
+struct CertifyTodoRequest: Encodable {
+    let content: String
+    let mediaUrls: [String]
+}
