+++ dogether/Data/Network/Request/SaveNotiTokenRequest.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 
-// TODO: 추후 수정
 struct SaveNotiTokenRequest: TokenInfo, Encodable {
-    let token: String = "kelly-token-value"
+    let token: String
 }
