+++ dogether/Data/Network/Response/AppleLoginResponse.swift
