+++ dogether/Data/Network/Response/GetGroupInfoResponse.swift
