+++ dogether/Data/Network/Response/GetMySummaryResponse.swift
