+++ dogether/Data/Protocol/GroupInfo.swift
