+++ dogether/Data/Protocol/ReviewInfo.swift
