+++ dogether/Data/Protocol/TokenInfo.swift
